Make ampel aware of the order in bearbeiter.txt
Technically this commit makes ampel way more fragile.
It now uses an array to store each employee and identifies
employees now by the index of this array. All this requires
everything to be in order (e.g. the dropdown in the login -
if this is out of order, then login results in another user
than the displayed one).
This was just a simple solution which works way to good.
An alternative whould be of cause another entry in the
previously used dict which shows an index.